backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0

python-dotenv==1.1.1

sqlalchemy==2.0.43
psycopg2-binary==2.9.10
asyncpg==0.29.0

google-generativeai==0.8.5
httpx==0.24.1
requests==2.32.5
pydantic==2.11.7
aiofiles==24.1.0
python-multipart==0.0.20
python-slugify==8.0.4

# JWT and cryptography for Clerk authentication
PyJWT==2.8.0
cryptography==41.0.7
supabase==1.2.0
websockets>=12.0

# Manim rendering (let manim resolve its own deps)
manim==0.19.0

# Manim API Documentation System
sentence-transformers==2.2.2
faiss-cpu>=1.7.4
beautifulsoup4==4.12.2
lxml==4.9.3

# Force CPU-only PyTorch to avoid large CUDA wheels in Cloud Run
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
torchvision==0.18.1+cpu

# Server
gunicorn==23.0.0 